+++ server/models/index.js
@@ -1,0 +1,8 @@
+const Lead = require('./Lead')
+const Mailbox = require('./Mailbox')
+const Template = require('./Template')
+const Campaign = require('./Campaign')
+const QueuedEmail = require('./QueuedEmail')
+const SendLog = require('./SendLog')
+
+module.exports = { Lead, Mailbox, Template, Campaign, QueuedEmail, SendLog }